feat: remove user route

+++ core/interfaces/UserRepository.go
@@ -6,7 +6,7 @@ import (
 
 type UserRepository interface {
 	Save(user models.User) error
-	Remove(id int) error
+	Remove(id int64) error
 	Put(id int, NewData models.User) error
 	FindOne(id int) (models.User, error)
 	Find() ([]models.User,error)
